requirements.txt
# Wymuszenie Python 3.11 (Python 3.13+ ma problemy z pydub)
python-version==3.11

openai
qdrant_client
pydub
streamlit
streamlit-audiorecorder
python-dotenv
moviepy<2.0
imageio==2.25.1
imageio-ffmpeg
boto3
psycopg2-binary
numpy
yt-dlp